add support for mp4 file renaming
Note that the current code requires GPS / location to be enabled in the
video in order to determine the timezone and hence determine the local
time that the video was taken

Note also that this code isn't very robust. I've only tested on some
video files from a Google Pixel phone. Other vendors could use different
formats for the metadata.